Title.java
CREATED Title class

CREATED splitPricingString to take the pricing info string from a file and converting it to a HashMap

Main.java

ADDED a test for the new Title class and it works correctly

+++ src/qc/ca/bdeb/_420_a02_bb/tp1/Order.java
@@ -4,7 +4,7 @@ public class Order {
     String title;
     int titleQty;
     String rebateName;
-    float price;
+    double price;
     public Order() {
 
     }
